AI_Tour_Guide/requirements.txt
altair==4.2.0
attrs==22.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.230105
Pillow==9.4.0
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==13.0.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.1
zipp==3.11.0
